NMS-10704: rename test and merge RestInfoIT
